R6.3: split vector/route into model/session/check
Facade mod.rs re-exports everything; no public path changes. The mutable
producer session (VectorRouteSession, transport declare/store/load
emission, chunk_index) lives in session.rs; the standalone checker
entries and the shared terminal verify path live in check.rs; region
layout, routed-trace DTOs, the VerifiedRoutedFir wrapper and the error
taxonomy in model.rs.

Plan §4.8 — admission guards in the shared verify function
`verify_routed_fir_with_policies_after_plan` (hosted in check.rs):
- verify_transport (declaration/store/load shape checks)
- verify_policy_coverage
- check_value_type / value_shape_matches

remain on BOTH paths after the split: the producer session's terminal
verification (session.rs finish path) and the standalone checker
entries (verify_routed_fir{,_with_clock_plan,_with_policies}). Existing
rejection tests fail through the standalone checker entry alone
(mutated traces -> TransportLoad, InvalidUse) and stay green.

Validation: transform 387 lib tests, transform clippy (+1.97.0) clean.
