+++ porting/journal/2026-07-20.md
@@ -72,3 +72,19 @@ modules on both sides of the plan artifact. Validation: transform 385,
   `lockstep_delay_one_register_mapping_is_checked_independently`) —
   inchangés et verts.
 - Gates : transform 387, clippy workspace vert, fmt vert.
+
+## R6.3 — scission de `vector/route`
+
+- `route/mod.rs` (1 428 lignes) scindé en `model.rs` (329), `session.rs`
+  (495, session mutable du producteur + émissions declare/store/load),
+  `check.rs` (567), façade `mod.rs` (38).
+- §4.8 : `verify_routed_fir_with_policies_after_plan` reste dans
+  `check.rs`, appelé par l'étape terminale de la session (producteur) ET
+  par les entrées checker autonomes (`verify_routed_fir`,
+  `verify_routed_fir_with_clock_plan`, `verify_routed_fir_with_policies`).
+  Tests de rejet par l'entrée checker seule déjà présents (traces mutées →
+  TransportLoad, InvalidUse) — inchangés et verts.
+- `value_fir_type` passait de `pub(super)` (= vector) à `pub(super)`
+  (= route) par le déplacement — élargi en
+  `pub(in crate::signal_fir::vector)` car `lower` l'importe.
+- Gates : transform 387, clippy 0, fmt vert.
